Connect submission links to target pages
Connected the forms to the pages the "Next" button previously had links
pointing to. Also made the WEBAUTHN link a button and hooked it up to a
new placeholder page using a JS onclick event listener.

It might be better to rename the file webauthn-redirect.js to something
like setupEventListeners.js so that it can be a generic event listener
setup file potentially used for multiple event listeners. However, it
might be better not to, because of the way event listeners are added
(add appropriate link to button with given ID) as it cannot be used
generically.