Close #39 - add confirmation action
Clicking the "remove security key" button triggers a 3-second interval
in which the user needs to press the button again to actually perform
the confirmation. This triggers a hidden button that submits the form.

The implications for UU are currently unexplored. This way of doing
things is subject to change.

Also updated instances of "sikkerhetsnøkkel" to conform with new naming
scheme of "smartenhet".