Close #105 - update device name in settings
Re-added the 'Avbryt'/'Lagre' buttons, and created an onclick listener
for the save button that emits a webauthn:update-device event. Added
event listener in localStorage.js that updates localStorage upon seeing
this event, and an event listener in update-localstorage-fields.js that
updates the fields on the page (not relevant to the settings page,
though).

Updating the device name in settings now works as expected. It's still
not possible to register a device from the page in question.