Add link to settings page from MinID innstillinger
Went through all pages and added a link to update-settings/index.html
from all pages containing a link with the text "MinID innstillinger"
(not always visible), in order to fix #128. Also made the settings index
'Tilbake' button link to document.referrer.

This is a WIP as the issue #132 still needs to be fixed in order for the
back button to work properly. The plan is to implement this and then
merge.